refactor(browser): separate executable probes and share classification
