+++ extensions/browser/src/browser-runtime.ts
@@ -53,11 +53,8 @@ export {
   DEFAULT_OPENCLAW_BROWSER_COLOR,
   DEFAULT_OPENCLAW_BROWSER_PROFILE_NAME,
 } from "./browser/constants.js";
-export {
-  parseBrowserMajorVersion,
-  readBrowserVersion,
-  resolveGoogleChromeExecutableForPlatform,
-} from "./browser/chrome.executables.js";
+export { parseBrowserMajorVersion, readBrowserVersion } from "./browser/chrome.executable-probe.js";
+export { resolveGoogleChromeExecutableForPlatform } from "./browser/chrome.executables.js";
 export { redactCdpUrl } from "./browser/cdp.helpers.js";
 export { DEFAULT_UPLOAD_DIR, resolveExistingPathsWithinRoot } from "./browser/paths.js";
 export { getBrowserProfileCapabilities } from "./browser/profile-capabilities.js";
