+++ extensions/browser/src/browser/chrome.test.ts
@@ -9,10 +9,8 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { WebSocketServer } from "ws";
 import { CHROME_STOP_PROBE_TIMEOUT_MS } from "./cdp-timeouts.js";
 import { diagnoseChromeCdp, formatChromeCdpDiagnostic } from "./chrome.diagnostics.js";
-import {
-  parseBrowserMajorVersion,
-  resolveGoogleChromeExecutableForPlatform,
-} from "./chrome.executables.js";
+import { parseBrowserMajorVersion } from "./chrome.executable-probe.js";
+import { resolveGoogleChromeExecutableForPlatform } from "./chrome.executables.js";
 import {
   getChromeWebSocketEndpoint,
   isChromeCdpOwnedByPid,
